fix: move to vertexai plugin
